disable `/bt friend`
