+++ tests/tower.test.ts
@@ -6,7 +6,13 @@ import { FakeTime } from '$std/testing/time.ts';
 
 import { assertSpyCalls, returnsNext, stub } from '$std/testing/mock.ts';
 
-import tower, { getFloorExp, MAX_FLOORS } from '~/src/tower.ts';
+import tower, {
+  getEnemyMaxSkillLevel,
+  getEnemyRating,
+  getEnemySkillSlots,
+  getFloorExp,
+  MAX_FLOORS,
+} from '~/src/tower.ts';
 
 import db from '~/db/mod.ts';
 import utils from '~/src/utils.ts';
@@ -59,6 +65,98 @@ Deno.test('exp gained amount floors 11-20', () => {
   assertEquals(getFloorExp(19), 3);
 
   assertEquals(getFloorExp(20), 6);
+});
+
+Deno.test('test enemy rating floors 1-10', () => {
+  assertEquals(getEnemyRating(1), 1);
+  assertEquals(getEnemyRating(2), 1);
+  assertEquals(getEnemyRating(3), 1);
+
+  assertEquals(getEnemyRating(4), 2);
+  assertEquals(getEnemyRating(6), 2);
+
+  assertEquals(getEnemyRating(7), 3);
+  assertEquals(getEnemyRating(8), 3);
+  assertEquals(getEnemyRating(9), 3);
+
+  assertEquals(getEnemyRating(5), 4);
+  assertEquals(getEnemyRating(10), 5);
+});
+
+Deno.test('test enemy rating floors 11-20', () => {
+  assertEquals(getEnemyRating(11), 1);
+  assertEquals(getEnemyRating(12), 1);
+  assertEquals(getEnemyRating(13), 1);
+
+  assertEquals(getEnemyRating(14), 2);
+  assertEquals(getEnemyRating(16), 2);
+
+  assertEquals(getEnemyRating(17), 3);
+  assertEquals(getEnemyRating(18), 3);
+  assertEquals(getEnemyRating(19), 3);
+
+  assertEquals(getEnemyRating(15), 4);
+  assertEquals(getEnemyRating(20), 5);
+});
+
+Deno.test('test enemy skill slots floors 1-10', () => {
+  assertEquals(getEnemySkillSlots(1), 0);
+  assertEquals(getEnemySkillSlots(2), 0);
+  assertEquals(getEnemySkillSlots(3), 0);
+  assertEquals(getEnemySkillSlots(4), 0);
+
+  assertEquals(getEnemySkillSlots(5), 0);
+  assertEquals(getEnemySkillSlots(6), 0);
+  assertEquals(getEnemySkillSlots(7), 0);
+  assertEquals(getEnemySkillSlots(8), 0);
+  assertEquals(getEnemySkillSlots(9), 0);
+
+  assertEquals(getEnemySkillSlots(10), 1);
+});
+
+Deno.test('test enemy skill slots floors 11-20', () => {
+  assertEquals(getEnemySkillSlots(11), 1);
+  assertEquals(getEnemySkillSlots(12), 1);
+  assertEquals(getEnemySkillSlots(13), 1);
+  assertEquals(getEnemySkillSlots(14), 1);
+
+  assertEquals(getEnemySkillSlots(15), 1);
+  assertEquals(getEnemySkillSlots(16), 1);
+  assertEquals(getEnemySkillSlots(17), 1);
+  assertEquals(getEnemySkillSlots(18), 1);
+  assertEquals(getEnemySkillSlots(19), 1);
+
+  assertEquals(getEnemySkillSlots(20), 2);
+});
+
+Deno.test('test enemy skill levels floors 1-10', () => {
+  assertEquals(getEnemyMaxSkillLevel(1), 1);
+  assertEquals(getEnemyMaxSkillLevel(2), 1);
+  assertEquals(getEnemyMaxSkillLevel(3), 1);
+  assertEquals(getEnemyMaxSkillLevel(4), 1);
+
+  assertEquals(getEnemyMaxSkillLevel(5), 1);
+  assertEquals(getEnemyMaxSkillLevel(6), 1);
+  assertEquals(getEnemyMaxSkillLevel(7), 1);
+  assertEquals(getEnemyMaxSkillLevel(8), 1);
+  assertEquals(getEnemyMaxSkillLevel(9), 1);
+
+  assertEquals(getEnemyMaxSkillLevel(10), 2);
+});
+
+Deno.test('test enemy skill levels floors 11-20', () => {
+  assertEquals(getEnemyMaxSkillLevel(11), 2);
+  assertEquals(getEnemyMaxSkillLevel(12), 2);
+  assertEquals(getEnemyMaxSkillLevel(13), 2);
+  assertEquals(getEnemyMaxSkillLevel(14), 2);
+
+  assertEquals(getEnemyMaxSkillLevel(15), 3);
+  assertEquals(getEnemyMaxSkillLevel(16), 3);
+  assertEquals(getEnemyMaxSkillLevel(17), 3);
+  assertEquals(getEnemyMaxSkillLevel(18), 3);
+  assertEquals(getEnemyMaxSkillLevel(19), 3);
+
+  assertEquals(getEnemyMaxSkillLevel(20), 4);
 });
 
 Deno.test('/tower view', async (test) => {
@@ -1346,6 +1444,18 @@ Deno.test('/sweep', async (test) => {
       () => undefined as any,
     );
 
+    const gainExpStub = stub(
+      db,
+      'gainExp',
+      () => ({
+        levelUp: 1,
+        skillPoints: 1,
+        statPoints: 3,
+        exp: 10,
+        expToLevel: 20,
+      }),
+    );
+
     const getInventoryStub = stub(
       db,
       'rechargeConsumables',
@@ -1410,15 +1520,294 @@ Deno.test('/sweep', async (test) => {
         }) as any,
     );
 
-    const t = {
-      set: () => t,
+    const atomicMock = {
+      set: () => atomicMock,
       commit: () => ({ ok: true }),
     };
 
     const atomicStub = stub(
       db.kv,
       'atomic',
-      () => t as any,
+      () => atomicMock as any,
+    );
+
+    config.combat = true;
+    config.appId = 'app_id';
+    config.origin = 'http://localhost:8000';
+
+    try {
+      const message = await tower.sweep({
+        token: 'test_token',
+        userId: 'user_id',
+        guildId: 'guild_id',
+      });
+
+      assertEquals(message.json(), {
+        type: 4,
+        data: {
+          attachments: [],
+          components: [],
+          embeds: [{
+            type: 'rich',
+            image: {
+              url: 'http://localhost:8000/assets/spinner3.gif',
+            },
+          }],
+        },
+      });
+
+      await timeStub.runMicrotasks();
+
+      assertSpyCalls(fetchStub, 2);
+
+      assertEquals(
+        fetchStub.calls[1].args[0],
+        'https://discord.com/api/v10/webhooks/app_id/test_token/messages/@original',
+      );
+
+      assertEquals(fetchStub.calls[1].args[1]?.method, 'PATCH');
+
+      assertEquals(
+        JSON.parse(
+          (fetchStub.calls[1].args[1]?.body as FormData)?.get(
+            'payload_json',
+          ) as any,
+        ),
+        {
+          attachments: [],
+          embeds: [
+            {
+              type: 'rich',
+              title: 'Floor 1',
+              description:
+                '**name 1** leveled up and gained 3 stat points and 1 skill point.\n' +
+                '**name 2** leveled up and gained 3 stat points and 1 skill point.\n' +
+                '**name 3** leveled up and gained 3 stat points and 1 skill point.\n' +
+                '**name 4** leveled up and gained 3 stat points and 1 skill point.\n' +
+                '**name 5** leveled up and gained 3 stat points and 1 skill point.',
+            },
+          ],
+          components: [
+            {
+              type: 1,
+              components: [
+                {
+                  custom_id: 'tsweep',
+                  label: '/sweep',
+                  style: 2,
+                  type: 2,
+                },
+              ],
+            },
+          ],
+        },
+      );
+    } finally {
+      delete config.appId;
+      delete config.origin;
+      delete config.combat;
+
+      timeStub.restore();
+      fetchStub.restore();
+
+      getGuildStub.restore();
+      getInstanceStub.restore();
+      getUserStub.restore();
+      getInventoryStub.restore();
+      getUserPartyStub.restore();
+      consumeSweepStub.restore();
+      gainExpStub.restore();
+      atomicStub.restore();
+    }
+  });
+
+  await test.step('level ups 2x', async () => {
+    const timeStub = new FakeTime();
+
+    const media: AniListMedia[] = [
+      {
+        id: '0',
+        type: MediaType.Anime,
+        title: {
+          english: 'title',
+        },
+      },
+    ];
+
+    const characters: AniListCharacter[] = [
+      {
+        id: '1',
+        name: {
+          full: 'name 1',
+        },
+      },
+      {
+        id: '2',
+        name: {
+          full: 'name 2',
+        },
+      },
+      {
+        id: '3',
+        name: {
+          full: 'name 3',
+        },
+      },
+      {
+        id: '4',
+        name: {
+          full: 'name 4',
+        },
+      },
+      {
+        id: '5',
+        name: {
+          full: 'name 5',
+        },
+      },
+    ];
+
+    const fetchStub = stub(
+      utils,
+      'fetchWithRetry',
+      returnsNext([
+        {
+          ok: true,
+          text: (() =>
+            Promise.resolve(JSON.stringify({
+              data: {
+                Page: {
+                  media,
+                  characters,
+                },
+              },
+            }))),
+        } as any,
+        {
+          ok: true,
+          text: (() =>
+            Promise.resolve(JSON.stringify({
+              data: {
+                Page: {
+                  media,
+                  characters,
+                },
+              },
+            }))),
+        } as any,
+        undefined,
+      ]),
+    );
+
+    const getUserStub = stub(
+      db,
+      'getUser',
+      () => 'user' as any,
+    );
+
+    const getGuildStub = stub(
+      db,
+      'getGuild',
+      () => 'guild' as any,
+    );
+
+    const getInstanceStub = stub(
+      db,
+      'getInstance',
+      () => 'instance' as any,
+    );
+
+    const consumeSweepStub = stub(
+      db,
+      'consumeSweep',
+      () => undefined as any,
+    );
+
+    const gainExpStub = stub(
+      db,
+      'gainExp',
+      () => ({
+        levelUp: 2,
+        skillPoints: 2,
+        statPoints: 6,
+        exp: 10,
+        expToLevel: 20,
+      }),
+    );
+
+    const getInventoryStub = stub(
+      db,
+      'rechargeConsumables',
+      () =>
+        ({
+          inventory: {
+            floorsCleared: 1,
+          },
+        }) as any,
+    );
+
+    const getUserPartyStub = stub(
+      db,
+      'getUserParty',
+      () =>
+        ({
+          member1: {
+            id: 'anilist:1',
+            mediaId: 'anilist:0',
+            rating: 1,
+            combat: {
+              level: 1,
+              exp: 29.5,
+            },
+          },
+          member2: {
+            id: 'anilist:2',
+            mediaId: 'anilist:0',
+            rating: 1,
+            combat: {
+              level: 1,
+              exp: 9.85,
+            },
+          },
+          member3: {
+            id: 'anilist:3',
+            mediaId: 'anilist:0',
+            rating: 1,
+            combat: {
+              level: 1,
+              exp: 9.85,
+            },
+          },
+          member4: {
+            id: 'anilist:4',
+            mediaId: 'anilist:0',
+            rating: 1,
+            combat: {
+              level: 1,
+              exp: 9.85,
+            },
+          },
+          member5: {
+            id: 'anilist:5',
+            mediaId: 'anilist:0',
+            rating: 1,
+            combat: {
+              level: 1,
+              exp: 9.85,
+            },
+          },
+        }) as any,
+    );
+
+    const atomicMock = {
+      set: () => atomicMock,
+      commit: () => ({ ok: true }),
+    };
+
+    const atomicStub = stub(
+      db.kv,
+      'atomic',
+      () => atomicMock as any,
     );
 
     config.combat = true;
@@ -1471,10 +1860,10 @@ Deno.test('/sweep', async (test) => {
               title: 'Floor 1',
               description:
                 '**name 1** leveled up 2x and gained 6 stat points and 2 skill points.\n' +
-                '**name 2** leveled up and gained 3 stat points and 1 skill point.\n' +
-                '**name 3** leveled up and gained 3 stat points and 1 skill point.\n' +
-                '**name 4** leveled up and gained 3 stat points and 1 skill point.\n' +
-                '**name 5** leveled up and gained 3 stat points and 1 skill point.',
+                '**name 2** leveled up 2x and gained 6 stat points and 2 skill points.\n' +
+                '**name 3** leveled up 2x and gained 6 stat points and 2 skill points.\n' +
+                '**name 4** leveled up 2x and gained 6 stat points and 2 skill points.\n' +
+                '**name 5** leveled up 2x and gained 6 stat points and 2 skill points.',
             },
           ],
           components: [
@@ -1506,6 +1895,7 @@ Deno.test('/sweep', async (test) => {
       getInventoryStub.restore();
       getUserPartyStub.restore();
       consumeSweepStub.restore();
+      gainExpStub.restore();
       atomicStub.restore();
     }
   });
